Move skit Toolchain globals into types/extensions [skip ci]
Remove Toolchain.swift, lifting its two file-scope functions:
- captureSwiftVersion() becomes a fileprivate method on Skit.Run
- syntaxKitCacheRoot() becomes ProcessInfo.syntaxKitCacheRoot(default:),
  with OutputCache supplying the platform default via a static
  defaultCacheRoot constant

OutputCache.init no longer throws (its only failure source is gone), so
the call site drops its try?. Update globals-audit.md to reflect the
removed file and lowered global-function count.
